+++ instrumentation/opentelemetry-instrumentation-grpc/src/opentelemetry/instrumentation/grpc/_server.py
@@ -315,7 +315,7 @@ class OpenTelemetryServerInterceptor(grpc.ServerInterceptor):
                             # we handle in our context wrapper.
                             # Here, we're interested in uncaught exceptions.
                             # pylint:disable=unidiomatic-typecheck
-                            if type(error) != Exception:
+                            if type(error) != Exception:  # noqa: E721
                                 span.record_exception(error)
                             raise error
 
@@ -342,6 +342,6 @@ class OpenTelemetryServerInterceptor(grpc.ServerInterceptor):
 
                 except Exception as error:
                     # pylint:disable=unidiomatic-typecheck
-                    if type(error) != Exception:
+                    if type(error) != Exception:  # noqa: E721
                         span.record_exception(error)
                     raise error